+++ temporian/core/event_set_ops.py
@@ -4090,8 +4090,8 @@ class EventSetOperations:
         self: EventSetOrNode,
         interval: Duration,
         align: bool = True,
-        include_right: bool = True,
-        include_left: bool = False,
+        after_last: bool = True,
+        before_first: bool = False,
     ) -> EventSetOrNode:
         """Generates timestamps at regular intervals in the range of a guide
         [`EventSet`][temporian.EventSet].
@@ -4102,7 +4102,7 @@ class EventSetOperations:
             >>> b = a.tick(interval=tp.duration.seconds(3), align=True)
             >>> b
             indexes: ...
-                    timestamps: [ 6. 9. 12. 15.]
+                    timestamps: [ 6. 9. 12. 15. 18.]
             ...
 
             ```
@@ -4113,19 +4113,40 @@ class EventSetOperations:
             >>> b = a.tick(interval=tp.duration.seconds(3), align=False)
             >>> b
             indexes: ...
-                    timestamps: [ 5. 8. 11. 14.]
+                    timestamps: [ 5. 8. 11. 14. 17.]
             ...
 
             ```
 
+        Example with before_first:
+            ```python
+            >>> a = tp.event_set(timestamps=[5, 9, 16])
+            >>> b = a.tick(interval=tp.duration.seconds(3), align=True, before_first=True)
+            >>> b
+            indexes: ...
+                    timestamps: [ 3. 6. 9. 12. 15. 18.]
+            ...
+
+            ```
+
+        Example without after_last:
+            ```python
+            >>> a = tp.event_set(timestamps=[5, 9, 16])
+            >>> b = a.tick(interval=tp.duration.seconds(3), align=True, after_last=False)
+            >>> b
+            indexes: ...
+                    timestamps: [ 6. 9. 12. 15.]
+            ...
+
+            ```
         Args:
             interval: Tick interval.
             align: If false, the first tick is generated at the first timestamp
                 (similar to [`EventSet.begin()`][temporian.EventSet.begin]).
                 If true (default), ticks are generated on timestamps that are
                 multiple of `interval`.
-            include_right: If True, a tick after the last timestamp is included.
-            include_left: If True, a tick before the first timestamp is included.
+            after_last: If True, a tick after the last timestamp is included.
+            before_first: If True, a tick before the first timestamp is included.
 
         Returns:
             A feature-less EventSet with regular timestamps.
@@ -4136,8 +4157,8 @@ class EventSetOperations:
             self,
             interval=interval,
             align=align,
-            include_right=include_right,
-            include_left=include_left,
+            after_last=after_last,
+            before_first=before_first,
         )
 
     def tick_calendar(
@@ -4148,8 +4169,8 @@ class EventSetOperations:
         mday: Optional[Union[int, Literal["*"]]] = None,
         month: Optional[Union[int, Literal["*"]]] = None,
         wday: Optional[Union[int, Literal["*"]]] = None,
-        include_right: bool = True,
-        include_left: bool = False,
+        after_last: bool = True,
+        before_first: bool = False,
     ) -> EventSetOrNode:
         """Generates events periodically at fixed times or dates e.g. each month.
 
@@ -4179,7 +4200,7 @@ class EventSetOperations:
             >>> b
             indexes: ...
             events:
-                (365 events):
+                (366 events):
                     timestamps: [...]
             ...
 
@@ -4189,7 +4210,7 @@ class EventSetOperations:
             >>> tp.glue(b.calendar_hour(), b.calendar_minute())
             indexes: ...
             events:
-                (365 events):
+                (366 events):
                     timestamps: [...]
                     'calendar_hour': [2 2 2 ... 2 2 2]
                     'calendar_minute': [30 30 30 ... 30 30 30]
@@ -4201,7 +4222,7 @@ class EventSetOperations:
             >>> b.calendar_day_of_month()
             indexes: ...
             events:
-                (12 events):
+                (13 events):
                     timestamps: [...]
                     'calendar_day_of_month': [5 5 5 ... 5 5 5]
             ...
@@ -4213,10 +4234,10 @@ class EventSetOperations:
             >>> tp.glue(b.calendar_day_of_month(), b.calendar_month())
             indexes: ...
             events:
-                (2 events):
+                (3 events):
                     timestamps: [...]
-                    'calendar_day_of_month': [1 1]
-                    'calendar_month': [2 2]
+                    'calendar_day_of_month': [1 1 1]
+                    'calendar_month': [2 2 2]
             ...
 
             >>> # Every second in the period  (2 hours -> 7200 seconds)
@@ -4237,7 +4258,7 @@ class EventSetOperations:
             >>> b
             indexes: ...
             events:
-                (120 events):
+                (121 events):
                     timestamps: [...]
             ...
 
@@ -4247,6 +4268,30 @@ class EventSetOperations:
                 ...
             ValueError: Can't set argument to None because previous and
             following arguments were specified. Set to '*' or an integer ...
+
+            >>> # not after_last
+            >>> a = tp.event_set(timestamps=["2020-02-01", "2020-04-01"])
+            >>> b = a.tick_calendar(mday=10, after_last=False)
+            >>> tp.glue(b.calendar_day_of_month(), b.calendar_month())
+            indexes: ...
+            events:
+                (2 events):
+                    timestamps: [...]
+                    'calendar_day_of_month': [10 10]
+                    'calendar_month': [2 3]
+            ...
+
+            >>> # before_first
+            >>> a = tp.event_set(timestamps=["2020-02-01", "2020-04-01"])
+            >>> b = a.tick_calendar(mday=10, before_first=True)
+            >>> tp.glue(b.calendar_day_of_month(), b.calendar_month())
+            indexes: ...
+            events:
+                (4 events):
+                    timestamps: [...]
+                    'calendar_day_of_month': [10 10 10 10]
+                    'calendar_month': [1 2 3 4]
+            ...
 
             ```
 
@@ -4266,8 +4311,13 @@ class EventSetOperations:
             wday: '*' (any day), None (auto) or number in range `[0-6]`
                     (Sun-Sat) to tick at particular day of week. Can only be
                     specified if `day_of_month` is `None`.
-            include_right: If True, a tick after the last timestamp is included.
-            include_left: If True, a tick before the first timestamp is included.
+            after_last: If True, a tick after the last timestamp is included.
+                    Useful for windows operations were you want the timestamps
+                    to be included in the range of the ticks.
+            before_first: If True, a tick before the first timestamp is
+                    included.
+                    Useful for windows operations were you want the timestamps
+                    to be included in the range of the ticks.
 
         Returns:
             A feature-less EventSet with timestamps at specified interval.
@@ -4282,8 +4332,8 @@ class EventSetOperations:
             mday=mday,
             month=month,
             wday=wday,
-            include_right=include_right,
-            include_left=include_left,
+            after_last=after_last,
+            before_first=before_first,
         )
 
     def timestamps(self: EventSetOrNode) -> EventSetOrNode: